Update breath_first_search
Breath_First_Search in which the root node is expanded first then all the successor of the root node are expanded next, f(n)-depth of the node, no. of actions it takes to reach the node. FIFO first in first out early goal test, BFS finds sol of minimal no. of action generation nodes of depth d.